+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixed
 - Do not inspect for unknown columns inside value array for `whereIn()`. #56
 - Do not inspect for unknown columns and tables inside array indexes. #44
+- Completion was not working on create/update methods called on relation. #41
 
 ### Updated
 - Bump platform version to 2021.2 and resolve new deprecations.